requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
transformers
accelerate
tqdm
selenium
webdriver-manager
bs4
requests
pandas
matplotlib
